+++ rook.py
@@ -1,4 +1,5 @@
 from pieces import Piece
+from moves import possible_positions_vd,possible_positions_va, possible_positions_hr, possible_positions_hl
 
 class Rook(Piece):
     white_str='♜'
@@ -7,58 +8,5 @@ class Rook(Piece):
     # TOTAL VALID POSITIONS WHERE TO MOVE A ROOK TO:
 
     def valid_positions(self,from_row,from_col):
-        possible_positions=(self.possible_positions_vd(from_row, from_col)+self.possible_positions_va(from_row, from_col)+self.possible_positions_hr(from_row, from_col)+self.possible_positions_hl(from_row, from_col))
+        possible_positions=(possible_positions_vd(self,from_row, from_col)+possible_positions_va(self,from_row, from_col)+possible_positions_hr(self,from_row, from_col)+possible_positions_hl(self,from_row, from_col))
         return possible_positions
-        
-    # POSSIBLE VERTICAL DESCENDANT POSITIONS TO MOVE A ROOK TO:
-
-    def possible_positions_vd(self,row,col):
-        possibles = []
-        for next_row in range(row + 1, 8):
-            other_piece = self.__board__.get_piece(next_row, col)
-            if other_piece is not None:
-                if other_piece.__color__ != self.__color__:
-                    possibles.append((next_row, col))  
-                break  
-            else:
-                possibles.append((next_row, col))  
-        return possibles
-
-    # POSSIBLE VERTICAL ASCENDANT POSITIONS TO MOVE A ROOK TO:
-
-    def possible_positions_va(self,row,col):
-        possibles=[]
-        for next_row in range(row-1,-1,-1):
-            other_piece = self.__board__.get_piece(next_row, col)
-            if other_piece is not None:
-                if other_piece.__color__ != self.__color__:
-                    possibles.append((next_row, col))  
-                break  
-            else:possibles.append((next_row,col))
-        return possibles
-
-    # POSSIBLE HORIZONTAL RIGHT POSITIONS TO MOVE A ROOK TO:
-
-    def possible_positions_hr(self,row,col):
-        possibles=[]
-        for next_col in range(col + 1, 8):
-            other_piece = self.__board__.get_piece(row,next_col)
-            if other_piece is not None:
-                if other_piece.__color__ != self.__color__:
-                    possibles.append((row,next_col))  
-                break  
-            else:possibles.append((row,next_col))
-        return possibles
-
-    # POSSIBLE HORIZONTAL LEFT POSITIONS TO MOVE A ROOK TO:
-
-    def possible_positions_hl(self,row,col):
-        possibles=[]
-        for next_col in range(col-1,-1,-1):
-            other_piece = self.__board__.get_piece(row,next_col)
-            if other_piece is not None:
-                if other_piece.__color__ != self.__color__:
-                    possibles.append((row,next_col))  
-                break  
-            else:possibles.append((row,next_col))
-        return possibles