+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.7] - 2024-09-27
+
+### Added
+
+- Moves file for all de moves that are in 2 or more pieces
+
+### Changed
+
+- Board str now has numbers from 0 to 7 not 1 to 8
+- Updated rook test because of the moves changes
+- Updated rook valid_positions method due to the moves changes
+
 ## [0.1.6] - 2024-09-26
 
 ### Added